:label: [#6281] Move type definitions around and update Pyright config
Now all of openforms.contrib.zgw can be type checked in CI.
